backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-dotenv